requirements.txt
Jinja2==2.6
argparse==1.2.1
bottle==0.11.6
python-mpd2==0.5.0
wsgiref==0.1.2
